+++ packages/shared/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./tenancy.ts";
 export * from "./dinero.ts";
+export * from "./barrio.ts";